pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-triage-env"
version = "1.0.0"
description = "Email Triage & Intelligent Routing OpenEnv Environment — Meta PyTorch OpenEnv Hackathon 2026"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["openenv", "email", "triage", "reinforcement-learning", "customer-support"]

dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "openai>=1.14.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.0",    # For FastAPI TestClient
]

[project.scripts]
# Start server:  uv run --project . server
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]

[tool.setuptools.package-dir]
"server" = "server"
